Move EvalString tests to eval_env_test.cc
