+++ configure.py
@@ -173,8 +173,8 @@ else:
 
 objs = []
 for name in ['build_test', 'build_log_test', 'disk_interface_test',
-             'graph_test', 'ninja_test', 'parsers_test', 'subprocess_test',
-             'util_test', 'clean_test', 'test']:
+             'eval_env_test', 'graph_test', 'ninja_test', 'parsers_test',
+             'subprocess_test', 'util_test', 'clean_test', 'test']:
     objs += cxx(name, variables=[('cflags', test_cflags)])
 
 if platform != 'mingw':
